+++ client/extensions/woocommerce/state/selectors/promotions.js
@@ -1,10 +1,15 @@
 /**
+ * External dependencies
+ */
+import { get } from 'lodash';
+
+/**
  * Internal dependencies
  */
 import { getSelectedSiteWithFallback } from '../sites/selectors';
 
 export function getPromotions( rootState, siteId = getSelectedSiteWithFallback( rootState ) ) {
-	const { promotions } = rootState.extensions.woocommerce.sites[ siteId ];
+	const { promotions } = get( rootState, [ 'extensions', 'woocommerce', 'sites', siteId ], {} );
 	return promotions.promotions;
 }
 
@@ -15,12 +20,12 @@ export function getPromotionsPage( rootState, siteId = getSelectedSiteWithFallba
 }
 
 export function getPromotionsCurrentPage( rootState ) {
-	const { promotions } = rootState.extensions.woocommerce.ui;
-	return promotions.currentPage;
+	const { list } = get( rootState, [ 'extensions', 'woocommerce', 'ui', 'promotions' ], {} );
+	return list.currentPage;
 }
 
 export function getPromotionsPerPage( rootState ) {
-	const { promotions } = rootState.extensions.woocommerce.ui;
-	return promotions.perPage;
+	const { list } = get( rootState, [ 'extensions', 'woocommerce', 'ui', 'promotions' ], {} );
+	return list.perPage;
 }
